requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
aiohttp==3.10.0
beautifulsoup4==4.12.3
certifi>=2024.0.0
python-dotenv==1.0.1
pytest==8.3.0
pytest-asyncio==0.24.0
